chore: Version Listener CRDs
